requirements.txt
altair==5.1.2
asttokens==2.4.0
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.9.2
nest-asyncio==1.5.8
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
pure-eval==0.2.2
pyarrow==13.0.0
pycodestyle==2.11.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.27.2
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
widgetsnbextension==4.0.9
zipp==3.17.0
